cleaned up a bit..
