+++ custom_components/weishaupt_modbus/entities.py
@@ -1,27 +1,16 @@
 """Build entitiy List and Update Coordinator."""
 
-import asyncio
-import logging
-import warnings
-
-from pymodbus import ModbusException
-
 from homeassistant.components.number import NumberEntity
 from homeassistant.components.select import SelectEntity
 from homeassistant.components.sensor import (
-    #    SensorDeviceClass,
     SensorEntity,
     SensorStateClass,
 )
 from homeassistant.const import CONF_PORT, CONF_PREFIX
 from homeassistant.core import callback
 from homeassistant.helpers.device_registry import DeviceInfo
-
-# from homeassistant.const import UnitOfEnergy, UnitOfTemperature
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-)
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import (
     CONST,
@@ -33,17 +22,15 @@ from .const import (
     CONF_HK4,
     CONF_HK5,
 )
-from .hpconst import DEVICES, TEMPRANGE_STD
+
 from .items import ModbusItem
+from .modbusobject import ModbusObject, ModbusAPI
 from .kennfeld import PowerMap
-from .modbusobject import ModbusObject
-
-logging.basicConfig()
-_LOGGER = logging.getLogger(__name__)
-_LOGGER.setLevel(logging.WARNING)
+from .coordinator import MyCoordinator
+from .hpconst import DEVICES
 
 
-async def check_available(modbus_item, config_entry) -> bool:
+async def check_available(modbus_item: ModbusItem, config_entry: ConfigEntry) -> bool:
     """function checks if item is valid and available"""
     if config_entry.data[CONF_HK2] is False:
         if modbus_item.device is DEVICES.HZ2:
@@ -69,7 +56,13 @@ async def check_available(modbus_item, config_entry) -> bool:
     return False
 
 
-async def build_entity_list(entries, config_entry, modbusitems, item_type, coordinator):
+async def build_entity_list(
+    entries,
+    config_entry,
+    modbusitems: ModbusItem,
+    item_type,
+    coordinator: MyCoordinator,
+):
     """Build entity list.
 
     function builds a list of entities that can be used as parameter by async_setup_entry()
@@ -121,126 +114,6 @@ async def build_entity_list(entries, config_entry, modbusitems, item_type, coord
     return entries
 
 
-class MyCoordinator(DataUpdateCoordinator):
-    """My custom coordinator."""
-
-    def __init__(self, hass, my_api, modbusitems):
-        """Initialize my coordinator."""
-        super().__init__(
-            hass,
-            _LOGGER,
-            # Name of the data. For logging purposes.
-            name="weishaupt-coordinator",
-            # Polling interval. Will only be polled if there are subscribers.
-            update_interval=CONST.SCAN_INTERVAL,
-            # Set always_update to `False` if the data returned from the
-            # api can be compared via `__eq__` to avoid duplicate updates
-            # being dispatched to listeners
-            always_update=True,
-        )
-        self._modbus_api = my_api
-        self._device = None  #: MyDevice | None = None
-        self._modbusitems = modbusitems
-        self._number_of_items = len(modbusitems)
-
-    async def get_value(self, modbus_item):
-        """Read a value from the modbus."""
-        mbo = ModbusObject(self._modbus_api, modbus_item)
-        if mbo is None:
-            modbus_item.state = None
-        modbus_item.state = await mbo.value
-        return modbus_item.state
-
-    async def get_value_a(self, modbus_item):
-        """Read a value from the modbus."""
-        mbo = ModbusObject(self._modbus_api, modbus_item)
-        if mbo is None:
-            return None
-        return await mbo.value
-
-    async def _async_setup(self):
-        """Set up the coordinator.
-
-        This is the place to set up your coordinator,
-        or to load data, that only needs to be loaded once.
-
-        This method will be called automatically during
-        coordinator.async_config_entry_first_refresh.
-        """
-        await self.fetch_data()
-
-    async def fetch_data(self, idx=None):
-        """Fetch all values from the modbus."""
-        # if idx is not None:
-        if idx is None:
-            # first run: Update all entitiys
-            to_update = tuple(range(len(self._modbusitems)))
-        elif len(idx) == 0:
-            # idx exists but is not yet filled up: Update all entitiys.
-            to_update = tuple(range(len(self._modbusitems)))
-        else:
-            # idx exists and is filled up: Update only entitys requested by the coordinator.
-            to_update = idx
-
-        # await self._modbus_api.connect()
-        for index in to_update:
-            item = self._modbusitems[index]
-            match item.type:
-                # here the entities are created with the parameters provided
-                # by the ModbusItem object
-                case TYPES.SENSOR | TYPES.NUMBER_RO | TYPES.NUMBER | TYPES.SELECT:
-                    await self.get_value(item)
-                case TYPES.SENSOR_CALC:
-                    r1 = await self.get_value_a(item)
-                    item_x = ModbusItem(
-                        item.getNumberFromText("x"),
-                        "x",
-                        FORMATS.TEMPERATUR,
-                        TYPES.SENSOR_CALC,
-                        DEVICES.SYS,
-                        TEMPRANGE_STD,
-                    )
-                    r2 = await self.get_value(item_x)
-                    item_y = ModbusItem(
-                        item.getNumberFromText("y"),
-                        "y",
-                        FORMATS.TEMPERATUR,
-                        TYPES.SENSOR_CALC,
-                        DEVICES.WP,
-                        TEMPRANGE_STD,
-                    )
-                    r3 = await self.get_value(item_y)
-
-                    item.state = [r1, r2, r3]
-
-    async def _async_update_data(self):
-        """Fetch data from API endpoint.
-
-        This is the place to pre-process the data to lookup tables
-        so entities can quickly look up their data.
-        """
-        # try:
-        # Note: asyncio.TimeoutError and aiohttp.ClientError are already
-        # handled by the data update coordinator.
-        async with asyncio.timeout(10):
-            # Grab active context variables to limit data required to be fetched from API
-            # Note: using context is not required if there is no need or ability to limit
-            # data retrieved from API.
-            try:
-                listening_idx = set(self.async_contexts())
-                return await self.fetch_data(listening_idx)
-            except ModbusException:
-                warnings.warn("connection to the heatpump failed")
-
-        # except:  # noqa: E722 # ApiAuthError as err:
-        #    return None
-        # Raising ConfigEntryAuthFailed will cancel future updates
-        # and start a config flow with SOURCE_REAUTH (async_step_reauth)
-        # raise ConfigEntryAuthFailed from err
-        # except ApiError as err:
-        #    raise UpdateFailed(f"Error communicating with API: {err}")
-
-
 class MyEntity:
     """An entity using CoordinatorEntity.
 
@@ -262,7 +135,9 @@ class MyEntity:
     _dev_device = ""
     _modbus_api = None
 
-    def __init__(self, config_entry, modbus_item, modbus_api) -> None:
+    def __init__(
+        self, config_entry, modbus_item: ModbusItem, modbus_api: ModbusAPI
+    ) -> None:
         """Initialize the entity."""
         self._config_entry = config_entry
         self._modbus_item = modbus_item
@@ -390,7 +265,13 @@ class MySensorEntity(CoordinatorEntity, SensorEntity, MyEntity):
     _attr_device_class = None
     _attr_state_class = None
 
-    def __init__(self, config_entry, modbus_item, coordinator, idx) -> None:
+    def __init__(
+        self,
+        config_entry: ConfigEntry,
+        modbus_item: ModbusItem,
+        coordinator: MyCoordinator,
+        idx,
+    ) -> None:
         super().__init__(coordinator, context=idx)
         self.idx = idx
         MyEntity.__init__(self, config_entry, modbus_item, coordinator._modbus_api)
@@ -400,10 +281,6 @@ class MySensorEntity(CoordinatorEntity, SensorEntity, MyEntity):
         """Handle updated data from the coordinator."""
         self._attr_native_value = self.translate_val(self._modbus_item.state)
         self.async_write_ha_state()
-
-    # async def async_update(self) -> None:
-    #    # the synching is done by the ModbusObject of the entity
-    #    self._attr_native_value = await self.translateVal
 
     @property
     def device_info(self) -> DeviceInfo:
@@ -421,7 +298,14 @@ class MyCalcSensorEntity(MySensorEntity):
     # calculates output from map
     my_map = None
 
-    def __init__(self, config_entry, modbus_item, coordinator, idx, pwrmap) -> None:
+    def __init__(
+        self,
+        config_entry: ConfigEntry,
+        modbus_item: ModbusItem,
+        coordinator: MyCoordinator,
+        idx,
+        pwrmap: PowerMap,
+    ) -> None:
         MySensorEntity.__init__(self, config_entry, modbus_item, coordinator, idx)
         self.my_map = pwrmap
 
@@ -480,7 +364,13 @@ class MyNumberEntity(CoordinatorEntity, NumberEntity, MyEntity):
     _attr_native_min_value = 10
     _attr_native_max_value = 60
 
-    def __init__(self, config_entry, modbus_item, coordinator, idx) -> None:
+    def __init__(
+        self,
+        config_entry: ConfigEntry,
+        modbus_item: ModbusItem,
+        coordinator: MyCoordinator,
+        idx,
+    ) -> None:
         super().__init__(coordinator, context=idx)
         self._idx = idx
         MyEntity.__init__(self, config_entry, modbus_item, coordinator._modbus_api)
@@ -516,7 +406,13 @@ class MySelectEntity(CoordinatorEntity, SelectEntity, MyEntity):
     options = []
     _attr_current_option = "FEHLER"
 
-    def __init__(self, config_entry, modbus_item, coordinator, idx) -> None:
+    def __init__(
+        self,
+        config_entry: ConfigEntry,
+        modbus_item: ModbusItem,
+        coordinator: MyCoordinator,
+        idx,
+    ) -> None:
         super().__init__(coordinator, context=idx)
         self._idx = idx
         MyEntity.__init__(self, config_entry, modbus_item, coordinator._modbus_api)
