+++ custom_components/weishaupt_modbus/number.py
@@ -5,33 +5,32 @@ from __future__ import annotations
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-# from homeassistant.helpers.typing import DiscoveryInfoType
 
-# from time import gmtime, strftime
 from .const import TYPES
-from .hpconst import ITEMLISTS
-from .entities import build_entity_list, MyCoordinator
-# from .modbusobject import ModbusAPI
+from .hpconst import DEVICELISTS
+from .entities import build_entity_list
+from .coordinator import MyCoordinator
 
 
 async def async_setup_entry(
     hass: HomeAssistant,
-    # config: ConfigType,
     config_entry: ConfigEntry,
     async_add_entities: AddEntitiesCallback,
-    # discovery_info: DiscoveryInfoType | None = None,
 ) -> None:
-    """Set up the sensor platform."""
+    """Set up the number platform."""
     _modbus_api = config_entry.runtime_data
 
+    # start with an empty list of entries
     entries = []
 
-    for _useless, item in enumerate(ITEMLISTS):
-        coordinator = MyCoordinator(hass, _modbus_api, item)
+    for _useless, device in enumerate(DEVICELISTS):
+        # we create one communicator per device and entry platform. This allows better scheduling than one
+        coordinator = MyCoordinator(hass, _modbus_api, device, config_entry)
+        #
         await coordinator.async_config_entry_first_refresh()
 
         entries = await build_entity_list(
-            entries, config_entry, item, TYPES.NUMBER, coordinator
+            entries, config_entry, device, TYPES.NUMBER, coordinator
         )
 
     async_add_entities(
